even more twitter stuff, though create a new tweet is still not working for some reason
git-svn-id: http://literalshore.com/svn/kickball/trunk@689 73bdbd85-dd75-0410-9d14-8527f93b1508

+++ Classes/MGTwitterEngine.h
@@ -95,7 +95,7 @@
 
 - (NSString *)getUpdate:(unsigned long)updateID; // statuses/show
 - (NSString *)sendUpdate:(NSString *)status; // statuses/update
-- (NSString *)sendUpdate:(NSString *)status inReplyTo:(unsigned long)updateID; // statuses/update
+- (NSString *)sendUpdate:(NSString *)status inReplyTo:(unsigned long long)updateID; // statuses/update
 
 - (NSString *)sendRetweet:(unsigned long long)updateID; // statuses/retweet
